requirements.txt
asgiref==3.6.0
black==23.1.0
boto3==1.26.84
botocore==1.29.84
click==8.1.3
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==20.1.0
isort==5.12.0
jmespath==1.0.1
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
